[metadata] rename module layout2 -> layout
