+++ core/src/storage2/alloc/allocator.rs
@@ -32,7 +32,7 @@ pub struct DynamicAllocator {
 #[cfg(feature = "std")]
 const _: () = {
     use crate::storage2::traits::StorageLayout;
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         FieldLayout,
         Layout,
         StructLayout,
