+++ core/src/storage2/alloc/boxed/storage.rs
@@ -30,7 +30,7 @@ use ink_primitives::Key;
 #[cfg(feature = "std")]
 const _: () = {
     use crate::storage2::traits::StorageLayout;
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         CellLayout,
         Layout,
         LayoutKey,
