+++ core/src/storage2/collections/hashmap/storage.rs
@@ -43,7 +43,7 @@ const _: () = {
             StorageLayout,
         },
     };
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         FieldLayout,
         Layout,
         StructLayout,
