+++ core/src/storage2/collections/smallvec/storage.rs
@@ -29,7 +29,7 @@ const _: () = {
         lazy::LazyArray,
         traits::StorageLayout,
     };
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         FieldLayout,
         Layout,
         StructLayout,
