+++ core/src/storage2/collections/vec/storage.rs
@@ -30,7 +30,7 @@ const _: () = {
         lazy::Lazy,
         traits::StorageLayout,
     };
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         FieldLayout,
         Layout,
         StructLayout,
