+++ core/src/storage2/lazy/lazy_array.rs
@@ -96,7 +96,7 @@ where
 #[cfg(feature = "std")]
 const _: () = {
     use crate::storage2::traits::StorageLayout;
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         ArrayLayout,
         CellLayout,
         Layout,
