+++ core/src/storage2/lazy/lazy_imap.rs
@@ -210,7 +210,7 @@ impl<V> LazyIndexMap<V> {
 #[cfg(feature = "std")]
 const _: () = {
     use crate::storage2::traits::StorageLayout;
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         ArrayLayout,
         CellLayout,
         Layout,
