+++ core/src/storage2/memory.rs
@@ -53,7 +53,7 @@ pub struct Memory<T> {
 #[cfg(feature = "std")]
 const _: () = {
     use crate::storage2::traits::StorageLayout;
-    use ink_metadata::layout2::{
+    use ink_metadata::layout::{
         CellLayout,
         Layout,
         LayoutKey,
